+++ test/modules/missing_receiver.rb
@@ -1,0 +1,2 @@
+export_from_receiver :Bar
+