+++ test/modules/receiver_sub.rb
@@ -1,0 +1,7 @@
+export_from_receiver :Sub
+
+class Sub < import('./receiver_super')
+  def self.bar
+    :bar
+  end
+end