+++ test/modules/receiver_super.rb
@@ -1,0 +1,7 @@
+export_default :Super
+
+class Super
+  def self.foo
+    :foo
+  end
+end